Fix the Dark Beam bug
sometimes a projectile would block the enemy, so the raycast was hitting the projectile. Since the raycast only checks on the first hit, and the first hit was the projectile, the dark beam would go through it and it would look like it didnt hit the enemy. Updated the prefabs for enemy and layer to be on their own layer, and updated the Raycast to only check colliders on the specific layers.