Add visual/sound for player hurt/death + update enemy hurt sound (#61)
Adding visual and sound effects for the player since there wasn't any feedback when the player was hurt. Luckily we did this for the enemy so it was an easy port.

I also swapped the enemy hurt sound to the player one, and gave the enemy a new one since that sound felt more right for the player.